requirements.txt
beautifulsoup4==4.12.2
build==1.0.3
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
extruct==0.16.0
html-text==0.5.2
html5lib==1.1
idna==3.4
ingredient-parser-nlp==0.1.0b5
isodate==0.6.1
joblib==1.3.2
jstyleson==0.0.2
lxml==4.9.3
mf2py==1.1.3
nltk==3.8.1
packaging==23.1
pyparsing==3.1.1
pyproject_hooks==1.0.0
pyRdfa3==3.5.3
python-crfsuite==0.9.9
rdflib==7.0.0
recipe-scrapers==14.49.0
regex==2023.8.8
requests==2.31.0
setuptools-scm==8.0.4
six==1.16.0
soupsieve==2.5
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==2.0.5
w3lib==2.1.2
webencodings==0.5.1
